src/app/counter-from-service/counter-from-service.service.ts
import { Injectable } from '@angular/core';
import { of } from 'rxjs';
import { delay, tap } from 'rxjs/operators';

@Injectable({
  providedIn: 'root'
})
export class CounterFromServiceService {

  currentCount = 0;

  increaseCount(count: 1 = 1): void {
    this.currentCount += count;
  }

  decreaseCount(count: 1 = 1): void {
    this.currentCount -= count;
  }

  getCount() {
    return of(1)
      .pipe(
        delay(1000),
        tap((value) => this.currentCount = value),
      );
  }
}
